Fix: job-allocation/resource cleanup race
This fixes a race condition where a job gets allocated, the block update
loop then makes a pass, and then starts to free the allocated job.  This can
result in the boot and the cleanup getting into a fight.

Additionally this catches an exception that was exposed by this race condition.