Boot functional and pybgsched mock library added
Issuing a standard initiate boot message starts the statemachine, and it
now progresses appropriate through a standard boot and reboot scenarion.
The code should be in proper condition to re-integrate back into the
bgqsystem code.

It should now be much easier to test using the mock pybgsched library,
adding this to the python path should allow it to be picked up
appropriately by all units.