Preeboot disable flag and bootable chilren client
Flag to qsub/qalter to request that a script job not be prebooted.  This
gets ignored for cN mode jobs.

Also adding in a client that lets us get the bootable blocks that are
children of a given block.  For use by scripts so they have an easier
time splitting up blocks for ensemble jobs.