Adding ION Boot control to pybgsched (0.2.0)
The bgsched python bindings are gaining ION control with these changes,
as well as the ability to boot single IONs and coherently pass recieved
exceptions/whether the callback was called back (IONBootCompletedFunc.h
is a container class that is wrapped for python's use in this).  This
verison requires V1R2M0 at a minimum to be used.