Fix for rapid block deletion bug and update loop failsafe
In the event of an error in update_block_state, all blocks will be set
to a "bridge-out" status.  Additionally an error in adding blocks has
been fixed that could result in an exception that can cause blocks to be
erroniously marked idle, if a block is added, cobalt detects it, and the
block is quickly deleted.