Fix for double-reservation entry
This was traced to a call that could cause a non-string key to be added
to the alps_reservation dictionary, resulting in a version of the
reservation with an integer jobid key and a second with a string jobid
key.  These should be keyed with strings.

Added as further mitigation a check to see if there is an integer
version of a key to clean.  If there is, then notify that it happened
and clean that one, too.

Triggering condition is an interactive job where the initial ALPS
reservation times out.