Test correction and dying now properly handled
The reservation would have gotten marked as dying even if the apids
hadn't been harvested.  Also, the data for the BASIL check was not
constructed correctly and while the right thing was happening, in that
an appication_id with a non-BASIL cmd was being resturned, the test
wasn't correct.  It's been fixed now.