Prevent the sytstem component from trying to start a second PG on retry
If we are retrying a process group startup, we needed to check to see if
there was an existing process group from a past call to
add_process_groups, where the execution happened, but the retrun failed,
or from another badly timed exception in the queue-manager.

If a process group has been added, this will retrieve the prior PG for a
job.  It will not invoke the startup a second time unless the job was
never actually started on the initial pass.