Tests no longer have time dependence.
In the original way these tests were written, we had a time-dependence
that would cause intermittent failures due to machine load that would
cause an error in the number of calls made over time, either the loop
would progress to far and go into a second iteration, or the first
iteration would not have completed by the time the asserts were checked.

This eliminates the timing dependence and ensures that we progress to
only one iteration-pass. If the tests hang, there is a timeout to
early-kill them with ample overhead to compensate for test-host lag or
high load.