Fix for a bad partial cleanup done if a job terminated during startup
This is the origin of a double allocate bug.  The behavior hinges on
this improper partial cleanup that occurs when a resource reservation
gets released during startup steps like job/resource prologues, or a
disrupted run.  By forcing this cleanup step an inconsistency between
the allocation timeout and running nodes is fixed.