Prevent the queue-manager from triggering an error on restart
If a job was in startup when Cobalt is shut down, we will have a
"starting" job, while the system component will have no relevant
allocation.  The job will fail and cleanup will be invoked, however,
there is no job data left to clean up from the system component.

Since there is no location or user data, we can't identify what needs
to be cleaned up at this point, so the function should skip that jobid.