Adding in reservation accounting log changes
This adds in reservation accounting log changes.  This is a WIP change,
and we're missing the reservation location data fetch for the Cray and
cluster systems.  I'm getting those changes into this branch next

Those changes are currently being extracted and are the next thing to be
integrated into this branch, however, isolating and reviewing these
changes seemed like a good idea.