Fixing a traceback from cyclic reservations
If the is_over check cycled the reservation, the running flag wasn't
getting set to False, causing a traceback to occur when the is_active
check happens later, as stime was set to None at this point (stime is
None in a reservation that has deactivated, and is only set while
active).

We would then get a loop where the reservation would never progress
because the log message would fail due to stime being set to None (which
it should be for a dormant reservation).