Start time estimate with unit tests and manpage update
This adds doc, unit tests and the starttime estimate.  This provides
a naive estimate based on machine hours ahead of a job, score-wise,
between the running jobs and currently eligible jobs.  Jobs in hold
do not get estimates, and putting a job in hold removes the estimate.

This also enforces a minimum interval ahead for start times.  This
is adjustable in the cobalt config.

This is a queue-data only estimate.  Cobalt is not prebuliding a
schedudle and the scheduler behavior is not impacted by this.

This will only show up by default with qstat -fl.  This may also
be requested via the header flag.  This is to prevent current user
script breakage.