Adding in reservation location finding tool
This adds the tool for finding a reservation location for cluster
systems.  This commit currently only supports cluster systems.

This will provide a list of hosts appropriate for feeding into the -p
flag of setres.  -t corresponds to duration and -s is the same as a
reservation start time.

--attrs won't work on cluster systems (the Cray should use it, though),
 but all the other flags should work.

Manpage pending.