Fix for double defer.  Simpler.
This needed to happen from is_over rather than is_active.  Not sure how
this got reversed.  Fixing a typo.  This actually causes the start time
to be adjusted on a normal cycle.

showres is actually smart enough to show correct times even if the start
time has not been set for the next cycle period on a natural end in the
prior code.