Lock doesn't need to be held around this state fetch
This is to improve Cobalt responsiveness during state update.  Combined
with the fix for aig/cobalt#177, this should allow for much faster
scheduling cadences as well and should significantly reduce the delay in
a number of commands.

There is an additional fix here where the check for a starting job
during the cleanup update wasn't working.  This was due to a hidden type
mismatch.  This was exposed by both speedups put together.