Fix for draining flap.
A node was getting added to the candidate list twice when looking over
currently running jobs. resulting in a duplicate node and an inflated
count which resulted in an early break.  When this list is later
converted to a set, the duplicates are removed and results in a set
of nodes too small for the job to use for draining, so the job doesn't
get drained for.

Test case added for the drain flap now works correctly now that a check
has been added so a duplicate node doesn't get added.